requirements.txt
Django==5.2.11
django-allauth==65.14.3
ruff==0.15.1
pre_commit==4.5.1
python-dotenv==1.2.2
asgiref==3.11.1
certifi==2026.2.25
cfgv==3.5.0
charset-normalizer==3.4.4
distlib==0.4.0
filelock==3.24.2
identify==2.6.16
idna==3.11
nodeenv==1.10.0
platformdirs==4.9.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
virtualenv==20.37.0
cffi==2.0.0
cryptography==46.0.5
jwt==1.4.0
pycparser==3.0